♻️ Refactor(character): Standardize schemas
