+++ backend/src/resources/character/player/dto/create-player.dto.ts
@@ -8,11 +8,11 @@ import {
 } from 'class-validator';
 import { ProgressionDto } from '@/resources/character/player/dto/progression/progression.dto';
 import { ClassDto } from '@/resources/character/player/dto/class/class.dto';
-import { ProfileDto } from '@/resources/character/dto/profile/profile.dto';
+import { PlayerProfileDto } from '@/resources/character/player/dto/profile/player-profile.dto';
 import { AppearanceDto } from '@/resources/character/player/dto/appearance/appearance.dto';
 import { BackgroundDto } from '@/resources/character/player/dto/background/background.dto';
 import { TreasureDto } from '@/resources/character/player/dto/treasure/treasure.dto';
-import { PlayerStatsDto } from '@/resources/character/dto/stats/stats.dto';
+import { PlayerStatsDto } from '@/resources/character/player/dto/stats/player-stats.dto';
 
 export class CreatePlayerDto extends CreateCharacterDto {
   @IsOptional()
@@ -32,8 +32,8 @@ export class CreatePlayerDto extends CreateCharacterDto {
 
   @ValidateNested()
   @IsOptional()
-  @Type(() => ProfileDto)
-  profile: ProfileDto;
+  @Type(() => PlayerProfileDto)
+  profile: PlayerProfileDto;
 
   @ValidateNested()
   @IsOptional()
