+++ backend/src/resources/character/player/dto/stats/player-stats.dto.ts
@@ -1,0 +1,8 @@
+import { IsNumber, IsOptional } from 'class-validator';
+import { StatsDto } from '@/resources/character/dto/stats/stats.dto';
+
+export class PlayerStatsDto extends StatsDto {
+  @IsOptional()
+  @IsNumber()
+  proficiencyBonus?: number;
+}
