+++ backend/src/resources/character/player/schemas/player.schema.ts
@@ -3,11 +3,11 @@ import { Document } from 'mongoose';
 import { Character } from '@/resources/character/schemas/character.schema';
 import { Progression } from '@/resources/character/player/schemas/progression/progression.schema';
 import { Class } from '@/resources/character/player/schemas/class/class.schema';
-import { PlayerProfile } from '@/resources/character/schemas/profile/profile.schema';
+import { PlayerProfile } from '@/resources/character/player/schemas/profile/player-profile.schema';
 import { Appearance } from '@/resources/character/player/schemas/appearance/appearance.schema';
 import { Background } from '@/resources/character/player/schemas/background/background.schema';
 import { Treasure } from '@/resources/character/player/schemas/treasure/treasure.schema';
-import { PlayerStats } from '@/resources/character/schemas/stats/stats.schema';
+import { PlayerStats } from '@/resources/character/player/schemas/stats/player-stats.schema';
 
 export type PlayerDocument = Player & Document;
