+++ backend/src/resources/character/player/schemas/profile/player-profile.schema.ts
@@ -1,0 +1,13 @@
+import { Profile } from '@/resources/character/schemas/profile/profile.schema';
+import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
+
+@Schema({ _id: false })
+export class PlayerProfile extends Profile {
+  @Prop()
+  race?: string;
+
+  @Prop()
+  subrace?: string;
+}
+
+export const PlayerProfileSchema = SchemaFactory.createForClass(PlayerProfile);
