+++ backend/src/resources/character/player/schemas/stats/player-stats.schema.ts
@@ -1,0 +1,10 @@
+import { Stats } from '@/resources/character/schemas/stats/stats.schema';
+import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
+
+@Schema({ _id: false })
+export class PlayerStats extends Stats {
+  @Prop({ default: 0 })
+  proficiencyBonus: number;
+}
+
+export const PlayerStatsSchema = SchemaFactory.createForClass(PlayerStats);
